# Batch series in streaming ingester based on message sizes.