@pstibrany 2020-08-13
Here is a result of benchmark before and after this change.

```
name                      old time/op    new time/op    deltanchstat before.txt after.txt 
Ingester_v2QueryStream-4    6.31ms ± 2%    6.46ms ± 1%  +2.38%  (p=0.001 n=10+8)

name                      old alloc/op   new alloc/op   delta
Ingester_v2QueryStream-4    3.45MB ± 0%    3.45MB ± 0%  +0.00%  (p=0.000 n=10+10)

name                      old allocs/op  new allocs/op  delta
Ingester_v2QueryStream-4     4.45k ± 0%     4.45k ± 0%  +0.02%  (p=0.000 n=10+10)
```