@pracucci 2020-08-13
Thanks @pstibrany! LGTM

> Should we do it for the chunks storage too, to keep it specular? :)

Could you open an issue for that, please?