@pstibrany 2020-08-13
> Could you open an issue for that, please?

I will take a look first if we can make it part of this PR (if it's small change).